move gauge instrumentation as part of initialization
