+++ .gitignore
@@ -17,3 +17,4 @@ Thumbs.db
 /.vercel/
 
 .idea
+/tmp/
