+++ pkg/execution/state/redis_state/queue.go
@@ -22,6 +22,7 @@ import (
 	osqueue "github.com/inngest/inngest/pkg/execution/queue"
 	"github.com/inngest/inngest/pkg/execution/state"
 	"github.com/inngest/inngest/pkg/logger"
+	"github.com/inngest/inngest/pkg/telemetry"
 	"github.com/oklog/ulid/v2"
 	"github.com/redis/rueidis"
 	"github.com/rs/zerolog"
@@ -212,6 +213,51 @@ func WithPollTick(t time.Duration) QueueOpt {
 func WithQueueItemIndexer(i QueueItemIndexer) QueueOpt {
 	return func(q *queue) {
 		q.itemIndexer = i
+	}
+}
+
+// WithAsyncInstrumentation registers all the async instrumentation that needs to happen on
+// each instrumentation cycle
+// These are mostly gauges for point in time metrics
+func WithAsyncInstrumentation(ctx context.Context) QueueOpt {
+	return func(q *queue) {
+		telemetry.GaugeWorkerQueueCapacity(ctx, q.capacity(), telemetry.GaugeOpt{PkgName: pkgName})
+
+		telemetry.GaugeGlobalQueuePartitionCount(ctx, telemetry.GaugeOpt{
+			PkgName: pkgName,
+			Observer: func(ctx context.Context) (int64, error) {
+				dur := time.Hour * 24 * 365
+				return q.partitionSize(ctx, q.kg.GlobalPartitionIndex(), getNow().Add(dur))
+			},
+		})
+
+		telemetry.GaugeGlobalQueuePartitionAvailable(ctx, telemetry.GaugeOpt{
+			PkgName: pkgName,
+			Observer: func(ctx context.Context) (int64, error) {
+				return q.partitionSize(ctx, q.kg.GlobalPartitionIndex(), getNow().Add(PartitionLookahead))
+			},
+		})
+
+		// Shard instrumentations
+		shards, err := q.getShards(ctx)
+		if err != nil {
+			q.logger.Error().Err(err).Msg("error retrieving shards")
+		}
+
+		telemetry.GaugeQueueShardCount(ctx, int64(len(shards)), telemetry.GaugeOpt{PkgName: pkgName})
+		for _, shard := range shards {
+			tags := map[string]any{"shard_name": shard.Name}
+
+			telemetry.GaugeQueueShardGuaranteedCapacityCount(ctx, int64(shard.GuaranteedCapacity), telemetry.GaugeOpt{PkgName: pkgName, Tags: tags})
+			telemetry.GaugeQueueShardLeaseCount(ctx, int64(len(shard.Leases)), telemetry.GaugeOpt{PkgName: pkgName, Tags: tags})
+			telemetry.GaugeQueueShardPartitionAvailableCount(ctx, telemetry.GaugeOpt{
+				PkgName: pkgName,
+				Tags:    tags,
+				Observer: func(ctx context.Context) (int64, error) {
+					return q.partitionSize(ctx, q.kg.ShardPartitionIndex(shard.Name), getNow().Add(PartitionLookahead))
+				},
+			})
+		}
 	}
 }
